requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
python-dotenv==1.0.0
twilio==8.10.0
apscheduler==3.10.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
